Fix setting up namespaces for mwxml
See https://github.com/mediawiki-utilities/python-mwxml/pull/9

I was relying on a hacked modification to mwxml, but now figured out that I can
solve the problem properly. I was trying to parse orphaned <page> elements from
the XML, but there is information in parent nodes that mwxml needs in order to
understand the <page> elements.

Now we just concatenate that information from the root node of the wiki dump.

So we can drop the submodule now, and install mwxml from pip directly.