HACK: Try to get around non-normalised names using SQL LIKE
Are all the names in the DB normalised? If so then I just have to normalise the
titles from wikilinks that I pass to get_seek_index. If not then... bit
trickier. Can't find a python lib to normalise titles (it's 3am though so
probably on exists) so for now just try this terrible hack.